Collapse activity labels of a sub process into a single activity

Collapse activity labels of a sub process into a single activity

act_collapse(log, ..., method, eventlog = deprecated())

# S3 method for eventlog
act_collapse(
  log,
  ...,
  method = c("entry_points", "consecutive"),
  eventlog = deprecated()
)

# S3 method for activitylog
act_collapse(
  log,
  ...,
  method = c("entry_points", "consecutive"),
  eventlog = deprecated()
)

# S3 method for grouped_log
act_collapse(
  log,
  ...,
  method = c("entry_points", "consecutive"),
  eventlog = deprecated()
)

Arguments

log

log: Object of class eventlog or activitylog.

...

A series of named character vectors. The activity labels in each vector will be collapsed into one activity with the name of the vector.

method

Defines how activities are collapsed: "entry_points" heuristically learns which of the specified activities occur at the start and end of the subprocess and collapses accordingly. "consecutive" collapses consecutive sequences of the activities.

eventlog

Deprecated; please use log instead.

Details

There are different strategies to collapse activity labels (argument ´method´). The "entry_points" method aims to learn the start and end activities of the sub process, by looking at the first and last activity in each case over the whole log. Subsequently, it will create a new instance of the sub process each time there is an end activity followed by a start activity. This strategy will not take into account other activities happening in the mean time. The "consecutive" method will create an instance each time a new sequence of sub activities is started. This strategy will thus only take into account interruptions of the other activity labels.

Methods (by class)

eventlog: Collapse activity labels of a subprocess into a single activity

activitylog: Collapse activity labels of a subprocess into a single activity

grouped_log: Collapse activity labels of a subprocess into a single activity

See also